refactor: Update npm dependencies and package versions
This commit updates the npm dependencies in package.json to their latest stable versions. Specifically, "@types/node" has been updated to version 20.14.0 and "vite" to version 5.2.12. Additionally, the package versions in package.json have been incremented from "1.0.6" to "1.0.7" in package-lock.json and from "1.0.6" to "1.0.7" in package.json. These changes improve the cleanliness, maintainability, and readability of the codebase.